requirements.txt
pandas
numpy
scikit-learn
scipy
PyWavelets
matplotlib
seaborn
gradio
langchain
langchain-core
langgraph
langchain-nvidia-ai-endpoints
langchain-tavily
tavily-python
fastapi
uvicorn
python-dotenv
joblib
tqdm
ipykernel